requirements.txt
# Core
pytz==2021.1
Django==3.2.7

# For the persistence stores
psycopg2-binary==2.9.1
dj-database-url==0.5.0


# Rest apis
djangorestframework==3.12.4
django-filter==2.4.0

# Developer Tools
ipdb==0.13.9
ipython==7.27.0
mkdocs==1.2.2
flake8==3.9.2

# Testing
mock==4.0.3
factory-boy==3.2.0
coverage==5.5


